Adding a command line parameter to convert references to ProjectReferences
Problem:
Normal Assembly references are not very smart, but ProjectReferences are smart.
A ProjectReference when used in a solution file will automatically determine build order.
Better yet a list of projects that are generated from an ItemGroup (in msbuild) and sent to a build task will automatically order themselves according to their ProjectReference. This is really cool. And thus a .sln file is almost completely unnecessary for a command line build environment.

This change allows you to point to a directory, and this tool will take all the msbuild projects, and convert their normal, week references to strong ProjectReference's.