Wrote a tool to find orphaned csharp files.
Given a directory, this will look for an examine all the csproj
files in it. For each csproj file it will look up all the code files
and store their name in a master list.
This will also, seperately look up all code files (*.cs) in the directory.

It will then find all the files that are not in a csproj file (Calling Except).

It then reports the number of orphaned files.