Add a more streamline way to convert to ProjectReferences
Problem:
The old code to convert References to ProjectReferences involved
a parameter for a search directory. It would then search for all the
projects in the search directory.
This was a useless and time consuming task that was utterly unnecessary.

Solution:
Refactor the SolutionBuilder to take just an msbuild file specifying exactly the files to operate on. This new method is almost instantaneous.