Lots of updates here:
Problem:
When a duplicate GUID was found in a project, the solution builder would incorrectly kick it out and
not include it. Even though it gave the project a new GUID.

Now it still gives it a new GUID and also includes it in the solution file.

Problem:
Csproj files that are .NET Core files use a simplified file format. This new simplified file format is so
radically different (and new) that Microsoft.MSBuild.dll that ships in the .NET framework does NOT
recognized the file format.
As a result, this solution builder would kick out these .NET Core project files, and the build would therefore
fail.

Fixed by using MSBuild API's from nuget.org. This has a nice advantage of this now being consistent across
all platforms, no matter what is installed.

Problem:
.NET Core csproj files generally do NOT have a GUID. Thus trying to add one to a solution file (*.sln) is problematic
and fails to load since the solution build just reported the GUID as blank. Well visual studio and MSBuild did NOT
like that and failed to load the projects.

I found that if I took a solution file and attempted to manually add a .NET Core project to it, that visual studio would simply
assign a GUID to the project and use that everywhere in the solution file (*.sln). It didn't necessarily modify the project, but
it all just works, since that GUID is only used in the sln file itself to determine build order. It's not used in actually compiling
the code.

Had to fix this by recognizing if a GUID was blank or empty and assigning it accordingly.

Problem 4
Some projects in a solution file do NOT have consistent platform targets. If you want to create a solution that targets AnyCPU,
there may be a project file that doesn't have that platform, but instead has 'x64'. As a result, if you load that solution, the project
gets loaded, and some properties are therefore not defined. This is usually solved by the user manually setting a different
configuration and platform for the project in the build configuration inside of visual studio.

This fixes that by checking if the project has the desired build platform defined. If the default platform is not the same as the desired
build platform, then it just uses the default platform the project defines and assigns that in the solution file.